A melange of nugatory fixes
- Fixed uninitialized variables in omitted.c and align.c.
- Avoid running a regression test in glob.sh that writes to '\',
  which is guaranteed to fail on Cygwin/Windows.
- ln(1) builtin: Don't treat -r as --recursive. This caused me
  grief in a script that was using ln -r with the assumption
  it was equivalent to GNU's --relative, which AST ln doesn't
  implement (yet).
- Backported a minor change from ksh93v- to swap md5sum's
  -T and -t options for compatibility with GNU coreutils.
- Override the native FTS library to fix a linking error
  when compiling on FreeBSD with clang and ASan.
- fastfind: Print a panic message in all out of memory scenarios.
- Since ksh is transitioning to C99, replace 'intmax_t' with
  'long long' where it's more appropriate from a purely pedantic
  standpoint.
- Denote some v1.1 changes in the NEWS file more clearly as
  belonging to v1.1 and not the v1.0 branch.
- Fixed many sign conversion warnings, particulary those involved
  in NOT operations with bitflags.
- sh_main(): Remove superfluous path_pwd() call. This call is
  pointless because path_pwd() will already have been called
  earlier in env_init(). If it failed there, it will fail here;
  if it succeeded, it'll be a no-op.
- Switch the O_cloexec and F_dupfd_cloexec definitions to
  uppercase (like in ksh93v-; I know this is purely stylistic but
  I quite frankly dislike seeing lowercase close-on-exec flags
  mixed with a bunch of uppercase flags).
- sh_open(): Switch from dup(2) to fcntl(2) because we need
  to preserve the close-on-exec flag if it was passed. Additionally,
  checking the flags needs to be done by masking those flags in
  case O_CLOEXEC was passed along with the flags.
- Correct the documentation for the astconflist() and
  sh_funscope() function arguments.
- strace was showing unnecessary fcntl calls in sh_pipe() even
  when SOCK_CLOEXEC is supported, so fix that to actually avoid
  those fcntl calls when it's completely superfluous.
- Delete unneeded _hdr_fcntl test for the standardized fcntl.h.
- Backport a minor correction for the description of [[ str > str2 ]]
  from the ksh93v- sh.1 manpage.

(upstreaming postponed due to dev branch quiescence)

+++ src/cmd/ksh93/bltins/cflow.c
@@ -67,7 +67,7 @@ done:
 	if(*argv)
 	{
 		int r;
-		intmax_t l = strtoll(*argv, NULL, 10);
+		long long l = strtoll(*argv, NULL, 10);
 		if(do_exit)
 		{
 			n = (int)(l & SH_EXITMASK);	/* exit: apply bitmask before conversion to avoid undefined int overflow */
